bot/requirements.txt
discord.py
wavelink
yt-dlp
python-dotenv
pycurl
